requirements.txt
asgiref==3.7.2
Django==4.2.1
django-debug-toolbar==4.1.0
Pillow==9.5.0
sqlparse==0.4.4
tzdata==2023.3
mysqlclient
python-dotenv
python-slugify
django-simple-captcha